app/geoprocessor/requirements.txt
celery[redis]
aiofiles
python-multipart
aiohttp
rasterio
shapely
geopandas
fiona
pillow